reorganize tooltip, data display, and utils for ecr-viewer (#2369)
* reorganizae tooltip and data display

* [pre-commit.ci] auto fixes from pre-commit hooks

* move utils.test.tsx

---------
