+++ containers/ecr-viewer/src/app/services/labsService.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Bundle, Device, Observation, Organization, Reference } from "fhir/r4";
-import { PathMappings, noData } from "@/app/utils";
+import { PathMappings, noData } from "@/app/view-data/utils/utils";
 import { evaluate } from "@/app/view-data/utils/evaluate";
 import { AccordionLabResults } from "@/app/view-data/components/AccordionLabResults";
 import {
@@ -16,7 +16,10 @@ import EvaluateTable, {
   ColumnInfoInput,
 } from "@/app/view-data/components/EvaluateTable";
 import { evaluateReference, evaluateValue } from "./evaluateFhirDataService";
-import { DataDisplay, DisplayDataProps } from "@/app/DataDisplay";
+import {
+  DataDisplay,
+  DisplayDataProps,
+} from "@/app/view-data/components/DataDisplay";
 
 export interface LabReport {
   result: Array<Reference>;
